+++ Editor/Core/ISettingsAnalyzer.cs
@@ -1,9 +1,8 @@
 using System;
 using System.Collections.Generic;
-using Unity.ProjectAuditor.Editor.Core;
 using UnityEditor;
 
-namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
+namespace Unity.ProjectAuditor.Editor.Core
 {
     public interface ISettingsAnalyzer
     {