+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Added 'documentation' support to descriptor
+* Added diagnostic area _Quality_, _Support_ and _Requirement_ 
+* Added Packages module to report installed packages and dependencies
+* Fixed analysis platform on incremental audit
+* Fixed compilation error due to newer com.unity.nuget.mono-cecil 
+* Improved text search to match custom properties
+
 ## [0.8.3-preview] - 2022-09-05
 * Fixed *NullReferenceException* on Draw2D shader not being found
 * Added HTML export support